requirements.txt
python-docx
xlrd
jieba
elasticsearch
